# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "واجهة برمجة التطبيقات لمنصة فاصلة للبودكاست",
        "title": "فاصلة API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "فريق دعم فاصلة",
            "url": "https://github.com/ziyadrw/faslah",
            "email": "zeadAlrouasheed@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/cms/create-content": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "إنشاء بودكاست جديد من خلال رفع ملف MP4 مباشرة أو من خلال رابط يوتيوب. يمكنك إما تقديم رابط فيديو يوتيوب وسنقوم بتنزيله واستخراج البيانات الوصفية تلقائيًا، أو يمكنك رفع ملف MP4 مباشرة وتقديم البيانات الوصفية يدويًا. في كلتا الحالتين، سيتم تخزين الفيديو في خدمة Cloudflare R2 الخاصة بنا تحت نطاق media.faslah.net للوصول السريع والآمن.",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "إنشاء المحتوى"
                ],
                "summary": "إنشاء بودكاست جديد",
                "parameters": [
                    {
                        "type": "string",
                        "description": "رابط يوتيوب (مطلوب إذا لم يتم تقديم ملف). سنقوم بتنزيل الفيديو واستخراج العنوان والوصف والوسوم تلقائيًا.",
                        "name": "source_url",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "ملف MP4 (مطلوب إذا لم يتم تقديم رابط يوتيوب). يجب أن يكون بتنسيق MP4.",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "العنوان (مطلوب لرفع الملف، اختياري لرابط يوتيوب). إذا قدمت رابط يوتيوب، يمكنك تجاوز العنوان المستخرج تلقائيًا.",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "الوصف (مطلوب لرفع الملف، اختياري لرابط يوتيوب). إذا قدمت رابط يوتيوب، يمكنك تجاوز الوصف المستخرج تلقائيًا.",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "الوسوم (مطلوبة لرفع الملف، اختيارية لرابط يوتيوب). قائمة من الكلمات المفتاحية لتسهيل البحث.",
                        "name": "tags",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "تاريخ النشر (اختياري). بتنسيق ISO 8601 (YYYY-MM-DDTHH:MM:SSZ).",
                        "name": "published_at",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم إنشاء البودكاست بنجاح",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_modules_cms_dtos.PodcastResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "خطأ في البيانات المدخلة",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "401": {
                        "description": "غير مصرح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "500": {
                        "description": "خطأ في الخادم",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    }
                }
            }
        },
        "/cms/delete-content/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "حذف بودكاست بواسطة المعرف",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "إدارة المحتوى"
                ],
                "summary": "حذف محتوى",
                "parameters": [
                    {
                        "type": "string",
                        "description": "معرف البودكاست",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم حذف البودكاست بنجاح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "400": {
                        "description": "معرف البودكاست غير صالح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "401": {
                        "description": "غير مصرح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "403": {
                        "description": "ليس لديك الصلاحيات الكافية",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "404": {
                        "description": "البودكاست غير موجود",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "500": {
                        "description": "خطأ في الخادم",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    }
                }
            }
        },
        "/cms/my-content": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "استرجاع جميع البودكاست التي أنشأها المستخدم الحالي",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "إدارة المحتوى"
                ],
                "summary": "الحصول على المحتوى الخاص بي",
                "responses": {
                    "200": {
                        "description": "تم استرجاع المحتوى الخاص بك بنجاح",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_modules_cms_dtos.PodcastResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "غير مصرح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "500": {
                        "description": "خطأ في الخادم",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    }
                }
            }
        },
        "/cms/retreive-content/{id}": {
            "get": {
                "description": "استرجاع بودكاست بواسطة المعرف",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "إدارة المحتوى"
                ],
                "summary": "الحصول على محتوى",
                "parameters": [
                    {
                        "type": "string",
                        "description": "معرف البودكاست",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم استرجاع البودكاست بنجاح",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_modules_cms_dtos.PodcastResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "معرف البودكاست غير صالح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "404": {
                        "description": "البودكاست غير موجود",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "500": {
                        "description": "خطأ في الخادم",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    }
                }
            }
        },
        "/cms/update-content/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "تحديث بودكاست بواسطة المعرف",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "إدارة المحتوى"
                ],
                "summary": "تحديث محتوى",
                "parameters": [
                    {
                        "type": "string",
                        "description": "معرف البودكاست",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "بيانات التحديث",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_modules_cms_dtos.UpdateContentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم تحديث البودكاست بنجاح",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_modules_cms_dtos.PodcastResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "خطأ في البيانات المدخلة",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "401": {
                        "description": "غير مصرح",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "403": {
                        "description": "ليس لديك الصلاحيات الكافية",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "404": {
                        "description": "البودكاست غير موجود",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    },
                    "500": {
                        "description": "خطأ في الخادم",
                        "schema": {
                            "$ref": "#/definitions/github_com_ziyadrw_faslah_internal_base.Response"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "أدخل رمز JWT مع البادئة Bearer، مثال: \"Bearer abcdef123456\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}